Refactored name and converted DeltaTime in sec
