+++ ScarletEditor/Sources/EditorMain.cpp
@@ -3,14 +3,14 @@
 #include "Editor.h"
 #include "RAL.h"
 #include "StaticMeshComponent.h"
-#include "Ac_RigidBodyComponent.h"
+#include "RigidBodyComponent.h"
 #include "AssetManager.h"
 
 using namespace ScarletEngine;
 
 void MakeCube(glm::vec3 Pos)
 {
-	auto [Ent, Trans, Mesh, Rb] = GEditor->GetActiveWorld()->CreateEntity<Transform, StaticMeshComponent, Ac_RigidBodyComponent>("Cube");
+	auto [Ent, Trans, Mesh, Rb] = GEditor->GetActiveWorld()->CreateEntity<Transform, StaticMeshComponent, Achilles::RigidBodyComponent>("Cube");
 
 	Trans->Position = Pos;
 	Trans->Rotation = glm::vec3(45.f, 45.f, 0.f);
