+++ ScarletEngine/CMakeLists.txt
@@ -76,4 +76,4 @@ add_subdirectory("Sources/OpenGLRAL")
 add_subdirectory("Sources/Renderer")
 add_subdirectory("Sources/RAL")
 add_subdirectory("Sources/Gameplay")
-add_subdirectory("Sources/Ac_Physics")
+add_subdirectory("Sources/AchillesPhysics")
